backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
redis>=4.3.0
python-multipart>=0.0.6
pypdf